+++ SmartDeviceLink/SDLDeleteWindow.h
@@ -15,7 +15,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  @param windowId A unique ID to identify the window. The value of '0' will always be the default main window on the main display and cannot be deleted.
  */
-- (instancetype)initWithId:(UInt32)windowId;
+- (instancetype)initWithId:(NSInteger)windowId;
 
 /**
  A unique ID to identify the window.
